Move schemas to separate file
